null pointer exception while deleting context caused by attempt to remove a non-existing context. As the apply of the log entry fails, all subsequent entries are not executed anymore. (#517)

+++ axonserver-enterprise/src/main/java/io/axoniq/axonserver/enterprise/context/ContextController.java
@@ -50,7 +50,9 @@ public class ContextController {
     public void updateContext(ContextConfiguration contextConfiguration) {
         Context context = entityManager.find(Context.class, contextConfiguration.getContext());
         if( ! contextConfiguration.getPending() && contextConfiguration.getNodesCount() == 0) {
-            entityManager.remove(context);
+            if (context != null) {
+                entityManager.remove(context);
+            }
             return;
         }
